example/backend/requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-filter==2.1.0
jsonfield==3.1.0

